requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1
django-ckeditor==6.4.2
django-dotenv==1.4.2
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-taggit==3.0.0
gunicorn==20.1.0
idna==3.3
Markdown==3.4.1
oauthlib==3.2.0
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
urllib3==1.26.11
whitenoise==6.2.0
